Fix tries counter.
Only decrement the tries counter when selecting an upstream server
that has not already been tried.

Decrementing the counter as part of the loop condition was prematurely
exiting the loop when the rehash hit on an upstream server which had
already been tried.

Ignore the hash_again parameter because now we'll try every peer in the
upstream block before failing.